+++ tpHeritage/Modele/PolyLigne.cpp
@@ -42,7 +42,7 @@ string PolyLigne::Obtenir_descripteur()
 	return a_retourner ;
 }
 
-bool PolyLigne::Deplacer(int dx, int dy, int id )
+bool PolyLigne::Deplacer(long int dx, long int dy, int id )
 {
 	return true ;
 }
